requirements.txt
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
Pillow==8.3.2
matplotlib==3.4.3
jupyter==1.0.0
